+++ amethyst_renderer/src/sprite.rs
@@ -107,7 +107,7 @@ impl Sprite {
         sprite_h: u32,
         pixel_left: u32,
         pixel_top: u32,
-        offsets: [i32; 2],
+        offsets: [f32; 2],
     ) -> Sprite {
         let image_w = image_w as f32;
         let image_h = image_h as f32;
@@ -204,13 +204,13 @@ impl Component for SpriteRender {
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 struct SpritePosition {
     /// Horizontal position of the sprite in the sprite sheet
-    pub x: f32,
+    pub x: u32,
     /// Vertical position of the sprite in the sprite sheet
-    pub y: f32,
+    pub y: u32,
     /// Width of the sprite
-    pub width: f32,
+    pub width: u32,
     /// Height of the sprite
-    pub height: f32,
+    pub height: u32,
     /// Number of pixels to shift the sprite to the left and down relative to the entity holding it
     pub offsets: Option<[f32; 2]>,
 }
@@ -219,9 +219,9 @@ struct SpritePosition {
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 struct SerializedSpriteSheet {
     /// Width of the sprite sheet
-    pub spritesheet_width: f32,
+    pub spritesheet_width: u32,
     /// Height of the sprite sheet
-    pub spritesheet_height: f32,
+    pub spritesheet_height: u32,
     /// Description of the sprites
     pub sprites: Vec<SpritePosition>,
 }
@@ -306,17 +306,16 @@ impl SimpleFormat<SpriteSheet> for SpriteSheetFormat {
         })?;
         let mut sprites: Vec<Sprite> = Vec::with_capacity(sheet.sprites.len());
         for sp in sheet.sprites {
-            sprites.push(Sprite {
-                width: sp.width,
-                height: sp.height,
-                offsets: sp.offsets.unwrap_or([0.0, 0.0]),
-                tex_coords: TextureCoordinates {
-                    left: sp.x / sheet.spritesheet_width,
-                    right: (sp.x + sp.width) / sheet.spritesheet_width,
-                    top: 1.0 - sp.y / sheet.spritesheet_height,
-                    bottom: 1.0 - (sp.y + sp.height) / sheet.spritesheet_height,
-                },
-            });
+            let sprite = Sprite::from_pixel_values(
+                sheet.spritesheet_width as u32,
+                sheet.spritesheet_height as u32,
+                sp.width as u32,
+                sp.height as u32,
+                sp.x as u32,
+                sp.y as u32,
+                sp.offsets.unwrap_or([0.0; 2]),
+            );
+            sprites.push(sprite);
         }
         Ok(SpriteSheet { texture, sprites })
     }
@@ -396,7 +395,7 @@ mod test {
         let sprite_h = 20;
         let pixel_left = 0;
         let pixel_top = 20;
-        let offsets = [-5, -10]; // Support negative offsets
+        let offsets = [-5.0, -10.0];
 
         assert_eq!(
             Sprite::from((